move action configs to separate file
